Safari-609.1.11.1/LayoutTests/editing/unsupported-content/script-markup-enabled-setting-expected.txt

This test verifies that script elements are not inserted into the document if the WebCore scriptMarkupEnabled setting is false.

PASS: documentWithScript.getElementsByTagName("script").length should be 0 and is.
PASS: documentWithSVGScript.getElementsByTagName("script").length should be 0 and is.
 
